fix: :bug: S3CardImage doesn't respect size boundary
