+++ apps/marketplace/src/components/marketplace/listing/ProductListingItem.tsx
@@ -17,12 +17,11 @@ import { useTheme } from '@mui/material/styles';
 import bookmarkListing from '@/middlewares/bookmarks/bookmarkListing';
 import { Listing } from '@/utils/api/client/zod';
 import { useSession } from 'next-auth/react';
-import S3Image from '@/components/S3Image';
+import S3CardImage from '@/components/S3CardImage';
 import MoreProfileIcon from './MoreProfileIcon';
 import BuyBadge from './BuyBadge';
 import SellBadge from './SellBadge';
 import NegotiableBadge from './NegotiableBadge';
-import CardMediaX from './CardMediaX';
 
 export type ProductListingItemData = {
   data: Listing;
@@ -95,7 +94,7 @@ const ProductListingItem = ({ data, showBookmark, updateBookmarkData }: ProductL
         />
       </Link>
       <Link style={{ textDecoration: 'none' }} href={`/listing/${listingName}-${data.id}`}>
-        <CardMediaX
+        <S3CardImage
           src={data.coverImage || placeholder}
           alt="listing image"
           height={200}
